refactor: split DropProgressToast into focused sub-250-line files
DropProgressToast.swift (436 lines) bundled four responsibilities. Split
verbatim into:

- DropCancellationToken.swift  - cancel token + DropCopyCancelled sentinel
- DropProgressCopier.swift     - chunked file copier
- DropProgressToast.swift      - toast state/coordination (lifecycle API)
- DropProgressToast+Panel.swift - AppKit panel construction/positioning

Each file is well under 250 lines and single-responsibility. The only
non-move change is widening the members the +Panel extension references
from private to internal; no behavior, signatures, or logic changed.
